+++ soccer/src/soccer/ui/field_view.cpp
@@ -405,7 +405,7 @@ void FieldView::drawTeamSpace(QPainter& p) {
     /// populate @cometTrails with the past locations of each robot
     int pastLocationCount = 40;  // number of past locations to show
     int start = std::max(0, static_cast<int>(_history->size()) - pastLocationCount);
-    for (int i = start; i < _history->size(); i++) {
+    for (unsigned long i = start; i < _history->size(); i++) {
         const LogFrame* oldFrame = _history->at(i).get();
         if (oldFrame != nullptr) {
             for (const LogFrame::Robot& r : oldFrame->self()) {
